feat: 라우터 변경.
커뮤니티, 스톡리스트,백테스트 ui통합
