+++ src/routes/main-router.tsx
@@ -5,9 +5,7 @@ import Portfolio from "./portfolio/page";
 import Dev from "./dev/page";
 import MainPortfolio from "./portfolio/mainPortfolio";
 import MyPage from "./mypage/page";
-import StockList from "./portfolio/stockList";
 import SignUp from "./signup/page";
-import CommPage from "./portfolio/comm_page";
 import SignIn from "./signin/page";
 import SubscribePortfolio from "./portfolio/subscribe/page";
 import DetailPage from "./portfolio/detail/page";
@@ -34,7 +32,7 @@ const routers = [
         index: true,
       },
       {
-        path: "detail/:accountId",
+        path: "detail/:id",
         element: <DetailPage />,
         index: true,
       },
@@ -46,16 +44,6 @@ const routers = [
       {
         path: "subscribe/recency",
         element: <SubscribePortfolioRecency />,
-        index: true,
-      },
-      {
-        path: "stockList/:id",
-        element: <StockList />,
-        index: true,
-      },
-      {
-        path: "comm",
-        element: <CommPage />,
         index: true,
       },
       {
