+++ src/routes/portfolio/detail/BackTest.tsx
@@ -1,0 +1,178 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { DatePicker } from "@mui/x-date-pickers/DatePicker";
+import { useEffect, useState } from "react";
+import dayjs, { Dayjs } from "dayjs";
+import { Slider } from "@mui/material";
+import { Pie } from "react-chartjs-2";
+import StockApi from "@/apis/stockAPI";
+
+type Props = {
+  id: string;
+};
+
+export default function BackTest({ id }: Props) {
+  const [startDate, setStartDate] = useState<Dayjs | null>(null);
+  const [endDate, setEndDate] = useState<Dayjs | null>(null);
+  const [accountdata, setAccountdata] = useState<any[]>([]); // stock_in_account 컬럼 값 저장
+  const [stockdata, setStockdata] = useState<any>({
+    // 파이차트에 대한 변수
+    labels: [],
+    datasets: [
+      {
+        label: "수량",
+        data: [],
+        backgroundColor: [
+          "rgba(255, 99, 132, 1)",
+          "rgba(54, 162, 235, 1)",
+          "rgba(255, 206, 86, 1)",
+          "rgba(75, 192, 192, 1)",
+          "rgba(153, 102, 255, 1)",
+          "rgba(255, 159, 64, 1)",
+        ],
+        borderWidth: 1,
+      },
+    ],
+  });
+
+  const today = dayjs();
+  const startYear2004 = dayjs("2004-01-01");
+
+  const ariav = (value: number) => {
+    return `date : ${value}`;
+  };
+
+  useEffect(() => {
+    const fetchAccountData = async () => {
+      if (!id) {
+        console.log("아이디없음", id);
+        return;
+      }
+      try {
+        const service = new StockApi();
+        console.log("Fetching account data...");
+        const account_res = await service.stockJoin(id);
+        const fetchedAccount = account_res;
+        console.log("Fetched account data:", fetchedAccount);
+
+        // 데이터 가공
+        const updatedData = fetchedAccount
+          .filter((account: any) => {
+            // console.log("Comparing", account.stock_id, "with", Number(id));
+            return account.account_id === Number(id);
+          })
+          .map((account: any) => {
+            const evlu_pfls_rt = (account.evlu_pfls_amt / account.pchs_amt) * 100;
+
+            return {
+              holdings_id: account.holdings_id,
+              account_id: account.account_id,
+              stock_id: account.stock_id,
+              market_id: account.market_id,
+              quantity: account.quantity,
+              pchs_amt: account.pchs_amt,
+              evlu_amt: account.evlu_amt,
+              evlu_pfls_amt: account.evlu_pfls_amt,
+              evlu_pfls_rt: evlu_pfls_rt,
+              stock_name: account.stock.name,
+              std_idst_clsf_cd_name: account.stock.std_idst_clsf_cd_name,
+            };
+          });
+        setAccountdata(updatedData);
+        // 파이차트
+        // 각 주식의 수량을 가져오는 코드
+        const quantities = updatedData.map((account: any) => account.quantity);
+
+        // 주식 이름을 찾는 코드
+        const stockNames = updatedData.map((account: any) => account.stock_name);
+        setStockdata({
+          labels: stockNames,
+          datasets: [
+            {
+              ...stockdata.datasets[0],
+              data: quantities,
+            },
+          ],
+        });
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      }
+    };
+
+    fetchAccountData();
+  }, [id]);
+  return (
+    <div className="portfolio-detail-container">
+      <div className="wrap-section flex flex-row">
+        <div className="section flex flex-col flex-1">
+          <div className="date-wrap flex">
+            <DatePicker
+              label={"시작"}
+              views={["month", "year"]}
+              value={startDate}
+              onChange={(newValue) => {
+                setStartDate(newValue);
+              }}
+              minDate={startYear2004}
+              maxDate={endDate || today} // endDate 이전 날짜만 선택 가능
+            />
+            <DatePicker
+              label={"종료"}
+              views={["month", "year"]}
+              value={endDate}
+              minDate={startDate || undefined} // startDate 이후 날짜만 선택 가능
+              maxDate={today} // endDate 이전 날짜만 선택 가능
+              onChange={(newValue) => {
+                setEndDate(newValue);
+              }}
+            />
+          </div>
+          {startDate && `${startDate.year()} ${startDate.month()}`}
+          <br />
+          {endDate && `${endDate.year()} ${endDate.month()}`}
+          <div className="range-wrap">
+            <Slider
+              aria-label="Temperature"
+              defaultValue={30}
+              getAriaValueText={ariav}
+              valueLabelDisplay="auto"
+              shiftStep={30}
+              step={10}
+              marks
+              min={10}
+              max={110}
+            />
+          </div>
+          <div className="chart-wrap">
+            <Pie
+              data={stockdata}
+              options={{
+                maintainAspectRatio: true,
+                responsive: false,
+                plugins: {
+                  legend: {
+                    display: false,
+                  },
+                },
+              }}
+            />
+            {accountdata.map((stock, i) => (
+              <div key={i} className="flex justify-between items-center mb-4 p-4 bg-gray-100 rounded-lg shadow">
+                <div className="text-left">
+                  <span className="block">{stock.stock_name}</span>
+                  <span className="block">{stock.quantity}주</span>
+                </div>
+                <div className="text-right">
+                  <span className="block">{stock.evlu_amt}원</span>
+                  <span className={`block ${stock.evlu_pfls_rt >= 0 ? "text-red-600" : "text-blue-600"}`}>
+                    {stock.evlu_pfls_amt}원<span>({parseFloat(stock.evlu_pfls_rt).toFixed(2)}%)</span>
+                  </span>
+                </div>
+              </div>
+            ))}
+          </div>
+        </div>
+        <div className="section flex-1">{/* chart */}</div>
+      </div>
+    </div>
+  );
+}
